Adding orbit_file-related files.
